Updated code for additional parameters
Added new middling classifier and reworked a lot of stuff involving success so that successful, non-successful and middling are now values for the bar charts, confusion matrix and pair plots. Also added a message for the Predictor that says whether the predicted GPA is successful, middling or not successful for both models.